serving/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
google-cloud-storage
pydantic
